requirements.txt
# Production dependencies
Flask>=2.3.0,<3.0.0
Werkzeug>=2.3.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
gunicorn>=21.2.0,<22.0.0

# Logging and monitoring
structlog>=23.1.0,<24.0.0
prometheus-flask-exporter>=0.22.3,<1.0.0

# Security
Flask-Limiter>=3.3.1,<4.0.0
Flask-Cors>=4.0.0,<5.0.0
flask-swagger-ui

# Development dependencies
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.11.1,<4.0.0
black>=23.7.0,<24.0.0
flake8>=6.1.0,<7.0.0
pre-commit>=3.3.3,<4.0.0
